# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StylistPlanRequestGender = typeof StylistPlanRequestGender[keyof typeof StylistPlanRequestGender];


export const StylistPlanRequestGender = {
  Women: 'Women',
  Men: 'Men',
  Unisex: 'Unisex',
} as const;

export type StylistPlanRequestSeason = typeof StylistPlanRequestSeason[keyof typeof StylistPlanRequestSeason];


export const StylistPlanRequestSeason = {
  Spring: 'Spring',
  Summer: 'Summer',
  Autumn: 'Autumn',
  Winter: 'Winter',
  All_Season: 'All Season',
} as const;

export interface StylistPlanRequest {
  gender: StylistPlanRequestGender;
  /** Casual, Date Night, Work, Vacation, Event, Streetwear, Evening, Party, Formal, Formal Remix, etc. */
  occasion: string;
  /** e.g. 'Under $500', '$500–$1500', '$1500–$3000', '$3000–$6000', '$6000+' */
  budget: string;
  season?: StylistPlanRequestSeason;
  /** @maxLength 500 */
  prompt?: string;
  /** @maxItems 10 */
  favoriteStyles?: string[];
}

export type StylistPlanSlotCategory = typeof StylistPlanSlotCategory[keyof typeof StylistPlanSlotCategory];


export const StylistPlanSlotCategory = {
  top: 'top',
  bottom: 'bottom',
  dress: 'dress',
  outerwear: 'outerwear',
  shoes: 'shoes',
  bag: 'bag',
  accessories: 'accessories',
  jewelry: 'jewelry',
} as const;

export type StylistPlanSlotFormality = typeof StylistPlanSlotFormality[keyof typeof StylistPlanSlotFormality];


export const StylistPlanSlotFormality = {
  casual: 'casual',
  smart: 'smart',
  dress: 'dress',
} as const;

export interface StylistPlanSlot {
  category: StylistPlanSlotCategory;
  /** One short editorial phrase, e.g. "fluid silk-jersey column dress" */
  descriptor: string;
  /** Preferred luxury houses for this slot, in priority order */
  brandPreferences?: string[];
  colorPreferences?: string[];
  formality?: StylistPlanSlotFormality;
}

export type StylistPlanResponseSeason = typeof StylistPlanResponseSeason[keyof typeof StylistPlanResponseSeason];


export const StylistPlanResponseSeason = {
  Spring: 'Spring',
  Summer: 'Summer',
  Autumn: 'Autumn',
  Winter: 'Winter',
  All_Season: 'All Season',
} as const;

export interface StylistPlanResponse {
  /** Dominant aesthetic, e.g. "Old Money", "Luxury Streetwear" */
  style: string;
  /** Palette name, e.g. "Champagne & Ivory" */
  palette: string;
  paletteColors: string[];
  season?: StylistPlanResponseSeason;
  /** Editorial look name, e.g. "Côte d'Azur Evening" */
  name: string;
  /** Two-sentence editorial blurb */
  description: string;
  /**
     * @minItems 2
     * @maxItems 6
     */
  slots: StylistPlanSlot[];
}

export type TierId = typeof TierId[keyof typeof TierId];


export const TierId = {
  basic: 'basic',
  premium: 'premium',
  pro: 'pro',
  vip: 'vip',
  diamond: 'diamond',
} as const;

export interface UsageTodayResponse {
  /** ISO date (YYYY-MM-DD) in server local time */
  date: string;
  /** @minimum 0 */
  looksGenerated: number;
  /**
     * 0 means uncapped (paid tiers report 0 here)
     * @minimum 0
     */
  capLimit: number;
  /** True when looksGenerated >= capLimit and capLimit > 0 */
  capped: boolean;
}

export interface LookAttemptRequest {
  /**
     * @minLength 1
     * @maxLength 128
     */
  userId: string;
  tier: TierId;
}

export interface LookAttemptResponse {
  allowed: boolean;
  /** @minimum 0 */
  looksGenerated: number;
  /** @minimum 0 */
  capLimit: number;
  /** Optional human-readable reason when allowed=false */
  reason?: string;
}

export type SubscriptionSyncRequestStatus = typeof SubscriptionSyncRequestStatus[keyof typeof SubscriptionSyncRequestStatus];


export const SubscriptionSyncRequestStatus = {
  active: 'active',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export interface SubscriptionSyncRequest {
  /**
     * @minLength 1
     * @maxLength 128
     */
  userId: string;
  tier: TierId;
  status: SubscriptionSyncRequestStatus;
}

export interface SubscriptionSyncResponse {
  ok: boolean;
  tier: TierId;
}

export type GetUsageTodayParams = {
/**
 * @minLength 1
 * @maxLength 128
 */
userId: string;
};