# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stylist
    description: AI stylist — generates structured outfit plans
  - name: usage
    description: Per-user daily usage metering (look generation caps for free tier)
  - name: subscriptions
    description: User subscription tier sync (mirrors RevenueCat entitlements server-side)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /stylist/plan:
    post:
      operationId: stylistPlan
      tags: [stylist]
      summary: Generate an AI stylist outfit plan
      description: |
        Returns a structured outfit plan (style, palette, slot specs) given
        the user's gender, occasion, budget, season, prompt, and favorite
        styles. The plan is consumed by the mobile outfit engine, which
        resolves each slot to a real catalog item while enforcing hard
        gender/season/aesthetic rules.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StylistPlanRequest"
      responses:
        "200":
          description: Plan generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StylistPlanResponse"
        "400":
          description: Invalid request body
        "502":
          description: Upstream AI provider error
  /usage/today:
    get:
      operationId: getUsageToday
      tags: [usage]
      summary: Get today's look-generation usage for a user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      responses:
        "200":
          description: Today's usage snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageTodayResponse"
        "400":
          description: Missing or invalid userId
  /usage/look-attempt:
    post:
      operationId: lookAttempt
      tags: [usage]
      summary: Record an attempted AI look generation, enforcing the free-tier daily cap
      description: |
        Atomically increments the per-user daily look counter ONLY when the
        attempt is allowed under the supplied tier. Free (basic) is capped at
        3/day. All paid tiers are uncapped. Returns whether the attempt was
        allowed plus the new counter state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LookAttemptRequest"
      responses:
        "200":
          description: Attempt resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookAttemptResponse"
        "400":
          description: Invalid request body
  /subscriptions/sync:
    post:
      operationId: syncSubscription
      tags: [subscriptions]
      summary: Mirror the user's current subscription tier server-side
      description: |
        Upserts the user's subscription tier so server-side gating (e.g. the
        free-tier look cap) can trust a single source of truth. Called by the
        client after every successful RevenueCat purchase, restore, or
        customerInfo refresh.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionSyncRequest"
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionSyncResponse"
        "400":
          description: Invalid request body
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    StylistPlanRequest:
      type: object
      properties:
        gender:
          type: string
          enum: [Women, Men, Unisex]
        occasion:
          type: string
          description: "Casual, Date Night, Work, Vacation, Event, Streetwear, Evening, Party, Formal, Formal Remix, etc."
        budget:
          type: string
          description: "e.g. 'Under $500', '$500–$1500', '$1500–$3000', '$3000–$6000', '$6000+'"
        season:
          type: string
          enum: [Spring, Summer, Autumn, Winter, "All Season"]
        prompt:
          type: string
          maxLength: 500
        favoriteStyles:
          type: array
          items:
            type: string
          maxItems: 10
      required:
        - gender
        - occasion
        - budget
    StylistPlanSlot:
      type: object
      properties:
        category:
          type: string
          enum: [top, bottom, dress, outerwear, shoes, bag, accessories, jewelry]
        descriptor:
          type: string
          description: One short editorial phrase, e.g. "fluid silk-jersey column dress"
        brandPreferences:
          type: array
          items:
            type: string
          description: Preferred luxury houses for this slot, in priority order
        colorPreferences:
          type: array
          items:
            type: string
        formality:
          type: string
          enum: [casual, smart, dress]
      required:
        - category
        - descriptor
    StylistPlanResponse:
      type: object
      properties:
        style:
          type: string
          description: Dominant aesthetic, e.g. "Old Money", "Luxury Streetwear"
        palette:
          type: string
          description: Palette name, e.g. "Champagne & Ivory"
        paletteColors:
          type: array
          items:
            type: string
        season:
          type: string
          enum: [Spring, Summer, Autumn, Winter, "All Season"]
        name:
          type: string
          description: Editorial look name, e.g. "Côte d'Azur Evening"
        description:
          type: string
          description: Two-sentence editorial blurb
        slots:
          type: array
          items:
            $ref: "#/components/schemas/StylistPlanSlot"
          minItems: 2
          maxItems: 6
      required:
        - style
        - palette
        - paletteColors
        - name
        - description
        - slots
    TierId:
      type: string
      enum: [basic, premium, pro, vip, diamond]
    UsageTodayResponse:
      type: object
      properties:
        date:
          type: string
          description: ISO date (YYYY-MM-DD) in server local time
        looksGenerated:
          type: integer
          minimum: 0
        capLimit:
          type: integer
          minimum: 0
          description: 0 means uncapped (paid tiers report 0 here)
        capped:
          type: boolean
          description: True when looksGenerated >= capLimit and capLimit > 0
      required:
        - date
        - looksGenerated
        - capLimit
        - capped
    LookAttemptRequest:
      type: object
      properties:
        userId:
          type: string
          minLength: 1
          maxLength: 128
        tier:
          $ref: "#/components/schemas/TierId"
      required:
        - userId
        - tier
    LookAttemptResponse:
      type: object
      properties:
        allowed:
          type: boolean
        looksGenerated:
          type: integer
          minimum: 0
        capLimit:
          type: integer
          minimum: 0
        reason:
          type: string
          description: Optional human-readable reason when allowed=false
      required:
        - allowed
        - looksGenerated
        - capLimit
    SubscriptionSyncRequest:
      type: object
      properties:
        userId:
          type: string
          minLength: 1
          maxLength: 128
        tier:
          $ref: "#/components/schemas/TierId"
        status:
          type: string
          enum: [active, expired, cancelled]
      required:
        - userId
        - tier
        - status
    SubscriptionSyncResponse:
      type: object
      properties:
        ok:
          type: boolean
        tier:
          $ref: "#/components/schemas/TierId"
      required:
        - ok
        - tier