# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Returns a structured outfit plan (style, palette, slot specs) given
the user's gender, occasion, budget, season, prompt, and favorite
styles. The plan is consumed by the mobile outfit engine, which
resolves each slot to a real catalog item while enforcing hard
gender/season/aesthetic rules.

 * @summary Generate an AI stylist outfit plan
 */
export const stylistPlanBodyPromptMax = 500;

export const stylistPlanBodyFavoriteStylesMax = 10;



export const StylistPlanBody = zod.object({
  "gender": zod.enum(['Women', 'Men', 'Unisex']),
  "occasion": zod.string().describe('Casual, Date Night, Work, Vacation, Event, Streetwear, Evening, Party, Formal, Formal Remix, etc.'),
  "budget": zod.string().describe('e.g. \'Under $500\', \'$500–$1500\', \'$1500–$3000\', \'$3000–$6000\', \'$6000+\''),
  "season": zod.enum(['Spring', 'Summer', 'Autumn', 'Winter', 'All Season']).optional(),
  "prompt": zod.string().max(stylistPlanBodyPromptMax).optional(),
  "favoriteStyles": zod.array(zod.string()).max(stylistPlanBodyFavoriteStylesMax).optional()
})

export const stylistPlanResponseSlotsMin = 2;
export const stylistPlanResponseSlotsMax = 6;



export const StylistPlanResponse = zod.object({
  "style": zod.string().describe('Dominant aesthetic, e.g. \"Old Money\", \"Luxury Streetwear\"'),
  "palette": zod.string().describe('Palette name, e.g. \"Champagne & Ivory\"'),
  "paletteColors": zod.array(zod.string()),
  "season": zod.enum(['Spring', 'Summer', 'Autumn', 'Winter', 'All Season']).optional(),
  "name": zod.string().describe('Editorial look name, e.g. \"Côte d\'Azur Evening\"'),
  "description": zod.string().describe('Two-sentence editorial blurb'),
  "slots": zod.array(zod.object({
  "category": zod.enum(['top', 'bottom', 'dress', 'outerwear', 'shoes', 'bag', 'accessories', 'jewelry']),
  "descriptor": zod.string().describe('One short editorial phrase, e.g. \"fluid silk-jersey column dress\"'),
  "brandPreferences": zod.array(zod.string()).optional().describe('Preferred luxury houses for this slot, in priority order'),
  "colorPreferences": zod.array(zod.string()).optional(),
  "formality": zod.enum(['casual', 'smart', 'dress']).optional()
})).min(stylistPlanResponseSlotsMin).max(stylistPlanResponseSlotsMax)
})


/**
 * @summary Get today's look-generation usage for a user
 */
export const getUsageTodayQueryUserIdMax = 128;



export const GetUsageTodayQueryParams = zod.object({
  "userId": zod.coerce.string().min(1).max(getUsageTodayQueryUserIdMax)
})

export const getUsageTodayResponseLooksGeneratedMin = 0;

export const getUsageTodayResponseCapLimitMin = 0;



export const GetUsageTodayResponse = zod.object({
  "date": zod.string().describe('ISO date (YYYY-MM-DD) in server local time'),
  "looksGenerated": zod.number().min(getUsageTodayResponseLooksGeneratedMin),
  "capLimit": zod.number().min(getUsageTodayResponseCapLimitMin).describe('0 means uncapped (paid tiers report 0 here)'),
  "capped": zod.boolean().describe('True when looksGenerated >= capLimit and capLimit > 0')
})


/**
 * Atomically increments the per-user daily look counter ONLY when the
attempt is allowed under the supplied tier. Free (basic) is capped at
3/day. All paid tiers are uncapped. Returns whether the attempt was
allowed plus the new counter state.

 * @summary Record an attempted AI look generation, enforcing the free-tier daily cap
 */
export const lookAttemptBodyUserIdMax = 128;



export const LookAttemptBody = zod.object({
  "userId": zod.string().min(1).max(lookAttemptBodyUserIdMax),
  "tier": zod.enum(['basic', 'premium', 'pro', 'vip', 'diamond'])
})

export const lookAttemptResponseLooksGeneratedMin = 0;

export const lookAttemptResponseCapLimitMin = 0;



export const LookAttemptResponse = zod.object({
  "allowed": zod.boolean(),
  "looksGenerated": zod.number().min(lookAttemptResponseLooksGeneratedMin),
  "capLimit": zod.number().min(lookAttemptResponseCapLimitMin),
  "reason": zod.string().optional().describe('Optional human-readable reason when allowed=false')
})


/**
 * Upserts the user's subscription tier so server-side gating (e.g. the
free-tier look cap) can trust a single source of truth. Called by the
client after every successful RevenueCat purchase, restore, or
customerInfo refresh.

 * @summary Mirror the user's current subscription tier server-side
 */
export const syncSubscriptionBodyUserIdMax = 128;



export const SyncSubscriptionBody = zod.object({
  "userId": zod.string().min(1).max(syncSubscriptionBodyUserIdMax),
  "tier": zod.enum(['basic', 'premium', 'pro', 'vip', 'diamond']),
  "status": zod.enum(['active', 'expired', 'cancelled'])
})

export const SyncSubscriptionResponse = zod.object({
  "ok": zod.boolean(),
  "tier": zod.enum(['basic', 'premium', 'pro', 'vip', 'diamond'])
})