# lib/api-zod/src/generated/types/lookAttemptRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TierId } from './tierId';

export interface LookAttemptRequest {
  /**
     * @minLength 1
     * @maxLength 128
     */
  userId: string;
  tier: TierId;
}