# lib/api-zod/src/generated/types/lookAttemptResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LookAttemptResponse {
  allowed: boolean;
  /** @minimum 0 */
  looksGenerated: number;
  /** @minimum 0 */
  capLimit: number;
  /** Optional human-readable reason when allowed=false */
  reason?: string;
}