# lib/api-zod/src/generated/types/subscriptionSyncResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TierId } from './tierId';

export interface SubscriptionSyncResponse {
  ok: boolean;
  tier: TierId;
}