# lib/api-zod/src/generated/types/tierId.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TierId = typeof TierId[keyof typeof TierId];


export const TierId = {
  basic: 'basic',
  premium: 'premium',
  pro: 'pro',
  vip: 'vip',
  diamond: 'diamond',
} as const;