# lib/api-zod/src/generated/types/usageTodayResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UsageTodayResponse {
  /** ISO date (YYYY-MM-DD) in server local time */
  date: string;
  /** @minimum 0 */
  looksGenerated: number;
  /**
     * 0 means uncapped (paid tiers report 0 here)
     * @minimum 0
     */
  capLimit: number;
  /** True when looksGenerated >= capLimit and capLimit > 0 */
  capped: boolean;
}